+++ config/config.go
@@ -15,7 +15,7 @@ import (
 	"github.com/spf13/viper"
 )
 
-const VERSION = "29.5"
+const VERSION = "29.6"
 
 type Config struct {
 	Server              ServerConfig